traderra/backend/requirements.txt
# Traderra Backend Dependencies
# AI and ML
pydantic-ai>=0.0.13
openai>=1.0.0
anthropic>=0.21.0

# FastAPI and Web Framework
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
python-multipart>=0.0.6

# Database
sqlalchemy>=2.0.0
asyncpg>=0.29.0
psycopg2-binary>=2.9.0

# Redis
redis>=5.0.0
aioredis>=2.0.0

# Authentication
python-jose[cryptography]>=3.3.0
python-multipart>=0.0.6
passlib[bcrypt]>=1.7.4

# HTTP Client for Archon MCP
httpx>=0.25.0
websockets>=12.0

# Environment and Configuration
python-dotenv>=1.0.0
pydantic-settings>=2.0.0

# Data Processing
pandas>=2.0.0
numpy>=1.24.0

# Utilities
python-dateutil>=2.8.0
pytz>=2023.3

# Development and Testing
pytest>=7.4.0
pytest-asyncio>=0.21.0
black>=23.0.0
isort>=5.12.0
mypy>=1.5.0

# Production
gunicorn>=21.2.0